Move IO functions to sub mod
We had everything in the top level, this change moves
the parse / compile implementations into a sub mod
which gives us better public / private separation
